requirements.txt
streamlit
langchain
langchain-openai
google-cloud-bigquery
langchain-google-genai
langchain_google_vertexai
langchain-community 
langchain-core
python-dotenv